% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{devBrain_geneSig_ref}
\alias{devBrain_geneSig_ref}
\title{Gene signatures for human developmental brain cell types.}
\format{
A data frame network with 112,018 rows and 3 columns:
}
\source{
\url{https://docs.google.com/spreadsheets/d/11E2_ENpnIettAYKPvWIA3VBopoHM4f1MlD1uhPbAyRc/edit?usp=sharing}
}
\usage{
devBrain_geneSig_ref
}
\description{
Gene signatures based on single-cell transcriptomic data reported by existing studies.
Gene symbols and their respective cell type or cluster annotation were extracted from original publications as reported and concatenated in one reference table.
Some cell types were relabeled for consistency across studies.
}
\keyword{datasets}
