% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testproj.R
\name{dweibull}
\alias{dweibull}
\title{Weibull Density}
\usage{
dweibull(t, scale, shape)
}
\arguments{
\item{t}{A vector of times}

\item{scale}{Scale parameter}

\item{shape}{Shape parameter}
}
\value{
The density
}
\description{
Weibull Density
}
\examples{
dweibull(1:10,1,3)
}

