% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testproj.R
\name{hweibull}
\alias{hweibull}
\title{Weibull Hazard Function}
\usage{
hweibull(t, scale, shape)
}
\arguments{
\item{t}{A vector of times}

\item{scale}{Scale parameter}

\item{shape}{Shape parameter}
}
\value{
Value of the hazard function
}
\description{
Weibull Hazard Function
}
\examples{
hweibull(1:10, 5,3)
}

