% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_poisson.R
\name{lasso_poisson}
\alias{lasso_poisson}
\title{Poisson regression with lasso regularity}
\usage{
lasso_poisson(X, y, lambda, max_iter = 300L, tol = 1e-07)
}
\arguments{
\item{X}{a matrix for predictors}

\item{y}{a non-negative integer vector}

\item{lambda}{a positive numeric vector or scalar.}

\item{max_iter}{the maximum iteration in coordiate descent algorithm}

\item{tol}{the tolorance for the coordinate descent algorithm}
}
\value{
a matrix of class \code{dgCMatrix}.
}
\description{
This function performs possion regression with lasso regularity.
}
\details{
The coordinate descent is used as in the package \code{glmnet}.
}
\examples{
x <- matrix(rnorm(1000), ncol = 5)
y <- rpois(200, 3)
fit <- lasso_poisson(x, y)

}
