% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testproj.R
\name{mweibull}
\alias{mweibull}
\title{Weibull Mean}
\usage{
mweibull(scale, shape)
}
\arguments{
\item{scale}{Scale parameter}

\item{shape}{Shape parameter}
}
\value{
The mean
}
\description{
Weibull Mean
}
\examples{
mweibull(1,3)
}

