% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_poisson.R
\name{plotx}
\alias{plotx}
\title{Plotting Poisson regression with lasso regularity}
\usage{
plotx(fit)
}
\arguments{
\item{fit}{an object which is an output from the function \code{lasso_poisson}.}
}
\description{
A function used to plot the Poisson regression with lasso regularity.
}
\examples{
x <- matrix(rnorm(1000), ncol = 5)
y <- rpois(200, 3)
fit <- lasso_poisson(x, y)
plotx(fit)
## the plot should be similar to the one in package glment
## fit2 <- glmnet(x, y)
## plot(fit2, "lambda")
}
