% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_poisson.R
\name{standardize}
\alias{standardize}
\title{Standardize a vector}
\usage{
standardize(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A numeric vector of the same length as input vector.
}
\description{
This function standardizes a numeric vector to have mean zero and variacne one.
}
\examples{
x <- rnorm(100)
y <- standardize(x)
mean(y)
sum(y^2)
}
