% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrepo-package.R
\docType{package}
\name{testrepo-package}
\alias{testrepo}
\alias{testrepo-package}
\title{testrepo: Test Repository With An R Package With One Or A Few Functions}
\description{
Test repository with an R package with one or a few functions.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
