% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_area.R
\name{triangle_area}
\alias{triangle_area}
\title{Compute the area of one or more triangles}
\usage{
triangle_area(base, height)
}
\arguments{
\item{base}{triangle base (numeric vector)}

\item{height}{triangle height (numeric vector)}
}
\description{
Compute the area of one or more triangles
}
\examples{
# Area of a triangle with base 1 and height 2
triangle_area(1, 2)

# Area of ten triangles, with base integers from 1 to 10 and height 2
triangle_area(1:10, 2)


}
