% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viral_update_gamma_john.R
\name{viral_update_gamma_john}
\alias{viral_update_gamma_john}
\title{Title}
\usage{
viral_update_gamma_john(dat, at)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
return value here.
}
\description{
Description
}
\details{
Additional details here
}
\examples{
example function call here
}
